feat: blockdevice domain with uuid persistence
In order to support matching of block devices with volume attachments
and to ensure referential integrity between block devices and volume
attachments, it is required to persist block device uuids between
block device refreshes by the diskmanager. This work implements the
usage of a de-duplication algorithm, this same algorithm is used to
match partial block devices as created by volume attachments or by
volume attachment plans with the block devices reported by the disk
manager.