refactor: state layer plumn for net node uuid and machine uuid
This commit fixes up the state layer for applications to now expect to
have been passed the net node uuid for new units and machine. As well as
this the machine uuid is expected to be passed to the state layer.

This puts an end to the state layer generating this information on
behalf of the caller. The reason for doing this is storage decisions
need to occur outside of the state layer and so all of the information
is required such as net node uuid.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>