refactor: storage instance ddl for charm ref
This commit removes the reference on storage_instance to a chamr storage
record. The reason for having done this is charms within the model
database are unique and not shared between applications.

This means that the current supported behaviour of Juju to allow
floating storage within the model (storage that is not attached to a
unit) can't be supported as it will always require the application and
charm to exist.

This would break juju attach storage and juju import storage. To combat
this modeling we are instead recording the charm name the storage
instance is intended to be used with.

This value will form as a check point for future code to make sure that
attachement of a storage instance to a new unit is inline with what our
expectations are.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>