feat: remove expirable storage
We never set the expire time for the underly storage to any value that
is different from the initial value. Ostensibly we were setting the
expiry multiple times assuming it could be configured, which wasn't
the case.

There are 3 types of bakeries:

 1. cross model bakery (local and JAAS)
 2. external macaroon authenticator for users (JAAS)
 3. local macaroon authenticator for users

1 is always set to three minutes, 2 and 3 are always set to 24 hours.

This is a big simplification, so should be double checked that my theory
is true.