feat: jaas bakery
This is the next in the series of patches that adds the JAAS bakery
back into the API server. This is mostly a lift from 3.6 but without
the ability to update the expiry time for a bakery.

The http client used for creating requests just used a http client
inline. We're tending to not want to do that anymore, instead we want
to get all http clients from a central worker. This is also updated
in this. By doing it this way, we don't need to mock out http round
trippers, instead just intercept the Do method of every client.