refactor: remove charm uuid from storage instance
This is a large commit that started out by just trying to remove usage
of charm_uuid from the storage_instance table. It turned into a small
refactoring of the tests within the state package of status.

It was noticed that there was an extreme double up of testing helpers in
this state package. These were coalesced into one base suite of helpers
that both storage suites can use.

This combination allowed for the removal of charm_uuid as well from the
helpers.

This commit also found two bugs in the current implementation that
exist. The first is we were using the associated charm storage kind to
inform the caller what the storage instance kind was. This is incorrect
as the mapping between charm storage and storage instance kind are not 1
to 1. It is internal buisness logic which determins this mapping.

The other bug found while here is around how status determines if a
filesystem or volume is provisioned within the model. The current code
assumes that this state is reached when the filesystem or volume is
associated with a storage instance. This is however not the case and the
current Juju logic we have determines this state via the fact that the
provisioning information has been set by the storage provisioning
worker. What we are most likely going to need moving forward is a
boolean flag indicating provisioned status or to put more stock in the
state machine that is status.

Both of these bugs have not been fixed in this commit and have only been
highlighted.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>