refactor: unique index on storage_attachment table
Recently we modified the storage_attachment table to include it's own
uuid as a primary key of the table. The reason for this was so that
custom triggers could be made to inform watchers in the model.

When doing this we removed the composite key that existed for
storage_instance_uuid and unit_uuid opting for a new unique index. The
new unique index was created with column order having unit_uuid first.
We already have an index on unit uuid and so what this essentially did
was remove the ability to look up attachments based on
storage_attachment_uuid quickly.

Any query that used this colume will result in a scan of the table. This
commit fixes this by putting the storage_instance_uuid column first in
the index.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>