refactor: comment out storage mount code
The storage domain had logic around how we generate mount locations for
filesystem attachments. This is specific provisioning logic that
directly relates to the storageprovisioning domain.

Specifically this is also needs to be integrated into the application
domain. I have commented this out for the moment so that a future PR can
reference and move the logic.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>