refactor: comment out storage attach code from application domain
This commit comments out the storage attach code that still lived in the
application domain state layer. This code is now broken as a result of
the storage_instance ddl change.

We already know that this code was not production ready to start with
and had errors. There is current work underway to integrate this
properly.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>