refactor(config): move config validation into the domain
Previously, validation of our application config on deploy was scattered
across the facade. However, this is business logic so not the concern on
the facade.

Move all this into the domain

Also, rename some types in internal/charm to match up with their
intention better. This is because "settings" has picked up a new meaning