fix(k8s): call destroy when NewModel fails for CAAS models
In this commit we add `broker.Destroy` when NewModel returns an error. This approach is consistent
to model's creation for iaas model. The reason for this change is that using Terraform we call
AddModel concurrently, this might fail with transaction aborted. With IAAS models we can just retry
on the transaction-aborted error, however with CAAS models we cannot because the k8s namespace
created with `broker.Create` is not cleaned up.