feat(operation): introduce export operation domain
Add functionality to handle exporting operations for model migration. This involves implementing
mechanisms to fetch and export operations and associated tasks.

The operation is not yet registered, it will be done as a follow up patch, once import and
underlying service method will be implemented.

- Added `exportOperation` struct and methods in `domain/operation/modelmigration/export.go`,
  including `Execute`, `Setup`, and helpers for task and operation handling.
- Introduced `ExportService` interface for operations retrieval.
- Implemented tests in `export_test.go` and `migrations_mock_test.go` to validate operation export
  functionality.
- Added `RegisterExport` method to register export operation with a coordinator.

This enables exporting operation metadata for model migration scenarios.