feat(config): refresh application config on charm refresh
The application config is an instantiation of the charm config. So when
a charm is refresh, the existing application config will potentially
need to be updated.

Handle this. Any existing application config, that can be coerced into
the new charm config, will be kept. Any config that can't will need to
be dropped.