feat: save external controller info during consume
When consuming an application offer, we need to save the external
controller information. This is only required if it's cross controller
relations.

This was already implemented, but ripped out when we had to remove all
the state calls. Thus, it was easy to put back, with a lot more tests.