refactor: model state tests for status
The model state tests for status needed to be refactored to now generate
uuids for machines and netnodes in application units.

While here it was found that a low of the tests were creating iaas units
for a caas application.

This commit tries to fix a lot of the testing mistakes but this file
still needs a lot of work.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>