fix(removal): correctly remove the net node on machine removal
Before this fix, the net node was never deleted, because of the delete node query:

```go
deleteNode := `
  DELETE FROM net_node WHERE uuid IN
  (SELECT net_node_uuid FROM machine WHERE uuid = $entityUUID.uuid)
```

This query was called after the machine deletion, so was always a noop.

The fix fetches the net node before removing the machine, then try to delete it.

A unit test has also been updated to trigger this error case.