test: check initial event using watcher harness
The watcher harness has a flaw which is currently being worked around,
which is that initial event checking is performed in a loop. If there
is more events than the initial event, it is confusing. The solution
here, is watch for only **one** event. This is the definition of the
initial event. You should only have one. Thus we should be able to
remove the work arounds that we have in place.