feat: add juju connected connection ips
In order to allow others to introspect the connected established
connections to the controller API server, offer out a common function.
This requires sudo, so we check for sudo before running, this isn't
the first time we done this, so we're walking over common ground.

Lastly, as the main API port can change, I've made it configurable along
with the total number of connections.

AWK is a bit weird with passing env vars into the script so we have to
setup a variable to pass it in.