refactor: create test caas application units with a net node uuid
This change refactors the application state test code that creates dummy caas units in the database to now make the units with a netnode uuid instead of an empty string.

This was missing from the original change of having the service layer generate unit net node uuids.

I tossed up with making the state layer funcs use sql.Null so that the DDL could cause an error on empty string but decided against it. At this stage the data must be right and I can't see a reason for the defensive programming at the moment.