feat: allow close server pending to give up
We can't wait forever for a request to go away. Give up and allow the
server to shed the connection. As the context isn't passed to every
method, unlike in 4.x, we can't assume that all method calls in state
will give up, so do our best.

In reality, if the connection has given up, I would expect that there
would be a slew of errors later on when what ever is holding the
connection attempts to write to the codec. That's fine and to be
expected. You had your chance.