feat: block units being added to cmr applications
Units are the manifestations of applications, in that they're the
units of the application replicated out. Thus if we prevent units
being added to CMR applications, we prevent any potential issues with
CMR synthetic applications. We've prevented CMR applications from
being located and returning the NotFound error, but we want to also
prevent units from ever being added. We can do this with a trigger that
just checks that it's never possible and an error will be thrown.