fix(add-model): fixes add-model for legacy facade versions
The transform from model qualifier to user tags for legacy add-model calls was broken for externals
users (those with a domain: e.g. alice@domain.com). In this PR I model the qualified creation into
the api package, once it is decide whether we are talking to an older controller or not.