feat: add check on same application for all actions on facade
This patch adds a check to make sure that all targets for the given
action are linked to the same application.

Also, it improves the result consolidation logic in the service layer as
well as the needed updates on state layer for this.
Tests now ensure proper behavior when having units that are the same as
a leader one.