refactor: storage kind in storage domain
I made the mistake of introducing another storage kind enum into the
storage provisioning domain when one already existed in the storage
domain.

This change fixes up the one in the storage domain so that it is better
aligned to it's purpose.

A follow PR will remove the storageprovisioning domain one.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>