feat(auth): enable service account login via env vars
This PR adds service account login (for JAAS) via the env vars JUJU_CLIENT_ID and
JUJU_CLIENT_SECRET, if they're present, they will take precendence over all other login providers.
Once logged in, the credentials are not persisted and expected to be drawn from the environm,ent
again on each command. It is a shame we require the svcacc to login, but there's no way around it as
we need the controller details.