fix: add a script for cleaning up obsolete secrets
We have had a few charms that have created secrets that did not
implement the secret-remove hook. This uses a database query to
determine the existing obsolete secret revisions, and then prints a list
of commands that can be run in the form of:
`juju exec --unit <unit> secret-remove <secret-id> --revision <x>` to
clean them all up.

This does not do the cleanup directly, but should make it reasonable to
actually try to clean up the 1000s of obsoleted secrets that we have on
our production controllers.