feat(refresh): wire up endpoint bindings refresh
An application's endpoint bindings have a foreign key to a specific
charm's relation. When an application is refreshed, we need to update
the endpoint bindings to point to the new charm.

This has a lot of subtleties, since the destination charm may have a
different set of endpoint bindings.

The simplest solution would have just been to have removed all
application endpoint and re-create them. However, other entities can
point to the endpoints meaning that we need to update them in place.

This isn't the case for extra endpoints, so we can just clear and
restore them.