feat: insert application endpoint bindings
Inserts the remote application offerer synthetic application endpoint
bindings. These are required when creating a relation. This is one
of the final steps before we can consume and relate applications on the
consumer side. Once this is done, we should be able to process the
changes and respond back to the offerer.

Although we pass the spaces through the parameters, they're unused. We
don't need to use them, instead all the application endpoints will
be placed into the alpha space (denoted by the NULL space id).