fix: if deleting multiple revisions don't delete everything
This is the same bug that was in my list_secrets script, as the code I
had written was inspired by Juju's code. If we ever actually deleted >1
secret revision, we would actually delete all secrets that had that
character in it.
eg, if deleting:
^f1837efb-828c-45c1-8e45-adc812f2b672:d3alij9fc4vpqb2sfg6g/(1)|(2)$

that actually deletes the thing we want (d3alij9.../1) and anything that
has a 2 in it.