fix: action result returned in the action facade must not be filled when error
This commit fixes a wrong behavior in the `toActionResult` method which
tries to fill the full result even when the TaskInfo contains an error.
In that case, the rest of the fields of TaskInfo are not guaranteed to
be non-nil, so this resulted in panics or empty fields.
The fix is to check whether we have an error in the task, and return
early in that case.