fix: enqueued operation result when nothing enqueued
This fixes a invalid tag error for the case where no valid targets were
passed, and therefore the service does not enqueue any task nor
operation.
In this case, the resulting operationID from the service is an empty
string, and the api layer must artificially add a tag (0) so that the
api client can parse it.