feat: rewrite operations query
This patch rewrites the complex and dynamic SQL query to retrieve
operations by filters.
The change is mainly that we need to apply pagination (limit + offset)
and therefore this limit has to be applied to the operation table only,
not to the JOINed results (which of course may include more rows than
the actual operations).
The solution is to apply the limit/offset within the WITH clause which
only retrieves the operation rows.