feat: get storage instances by provider id
This commit introduces to the application state layer the ability to get
storage instances in the model based on the provisioned provider id.
This is the provider id that is assocated with either the filesystem or
volume.

The finding method also returns the composition of the storage so that
the caller understands what is available. This state layer func
specifically exists to serve caas unit registration.

Co-authored-by: Thomas Miller <thomas@tlm.id.au>
Signed-off-by: Thomas Miller <thomas.miller@canonical.com>