fix(juju-qa-test): fix juju-qa-test charm to avoid overriding status
right now when we change `config.foo-file` the _on_config_changed
is updating the status message taking the first line of the file.
However it then overrides the status coming from config.status.
This doesn't happen in Juju 3.6 because this second update is lost,
but in Juju 4 we catch both of them.
This is causing error in all of the test where we wait for the status
message to be a particular string coming from the file.

The fix in this PR  basically avoids the update coming from config.status
if the status has been updated from the foo-file config.