fix(arch): arch is no longer required for charmhub
We recently made some slight chanegs to how arch was handled. An arch
used to be requried to call the charmhub API code. This cause problems,
since we lost whether the arch was a set constraint, or just a default
value which could be overriden by the return of charmhub.

However, a number of checks were still in place that asserted an arch
was provided when it was no longer needed.

This commit drops these checks where they are no longer needed.