refactor: representation of a storage directive in the application domain
I have refactored a little storage directives in the application domain
to now include the maximum storage count supported by the charm for a
directive.

The reason for this is the directive tells us the desired amount but it
is possible and reasonable for a single unit to be over this based on
user actions in the model.

When re-attaching storage for a unit we need to consider all of the
storage that must be re-attached and not what it desired.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>