refactor(config): we always store application config as strings in state
Make use of this to specify our type down from any, to sql.Null[string].

This means we return config values always as *string from the state
layer. This is incorrect, but not actually a change to behaviour.

The next step will be to coerce config values to the correct types in
the service layer.