refactor(application): refactor application service and type to use more solid types
Now application service talks with the state using two types, both of them are string's type, and
the reason is to avoid issues around letting the state layer converting back and forth from nil. In
addition to that, having two types help us to handle the case where we want to express that Getting
the config can yield nil value, but setting the config is always done via string values.