feat: moves storage related services to new package
Because of the sheer size of the application service package it is
becoming increasingly hard to manage the package and test the package.
There is also a lot of cross cutting concerns that are getting mixed up.

It is there fore easier to move storage related code out into another
package providing encapsulation and also making tests of application
code easier.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>