refactor: isolate application storage service code
It has become to hard to maintain the application service domain. There
are too many seperate concerns bleeding into each other. Because of this
I have broken out the storage logic into a seperate package.

This allows for testing to mock the storage service and for stricter
isolation of code.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>