fix(operation): exec task should have default "juju-exec" name
Set a default action name `JujuExecActionName` for tasks with no provided name in
 `domain/operation/state/task.go`.

This is necessary to get the right spec in `internal/worker/machineactions/handleactions.go#27` or
check that it is a "JujuExecAction" in various situation.

The action name is defaulted in the state while retrieving the action, since the value is not
persisted. Another approach would have been to use an empty string to reflect this special action,
but it would have required to hunt down all logic which rely on the special string `juju-exec`,
which is error prone.