feat: return relation uuid instead of relation key
The change allows us to return the relation UUID instead of the relation
key. We still want the prevention of sending duplicate messages, but
we don't want the relation key. The solution is to make the supporting
watcher a generic type of T, over the union of relation UUID and Key.
Thus we can have our cake and eat it.