fix(application-config): application config query to avoid sharing config between apps
Before the query was problematic because since the JOIN was done based of the charm's key if two
applications had the same charm uuid, they were sharing the application config. In this PR I've
fixed it, and created some utils function to make it easier to add applications with the same
charm's reference.