fix: normal rpc conn closures should not error on final close
During the handling of rpc connection closure, it is possible to
determine a normal closure.

Due to the reciprocal nature of normal closure, the rpc connection
may already be closed, this change ignores errors writing the
control message to close the connection if it is already closed.
It now actually reads the reciprocal close on success (rather than
failure), in addition does not send a reciprocal close during the
reading of the expected close control message.

Finally due to the early close of the tls/tcp connection to writing,
it is expected that the underlying tls connection may error during
the final close. Since it is possible to determine a normal closure
of the rpc connection, it is safe to ignore certain errors during
the final tls/tcp connection close.