refactor: making of register caas storage args
This change functionally seperates making register caas unit storage
args into a variant for existing units and new units. Before this was
done through introspection in the function that was awkward and didn't
take advantage of knowledge that the caller already had.

This new way lets the caller make the decision and in turn makes the
generations of args more procedural.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>