refactor: register caas unit
This commit calls the new make storage funcs in the storage service
depending on if the unit exists or does not exist. This code is now much
more easier to reason about and less error prone and repetitive.

I also remove a Println that should not have been in here.

Further to the above the code does not try to generate a unit uuid for
new units any more. Instead this has been replaced with a TODO.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>