feat(action): add support for leader units in operations
Introduce handling for leader units in operations by updating `makeOperationReceivers` in
`apiserver/facades/client/action/action.go`. Units with a `/leader` suffix are now identified,
separated, and stored as a new `LeaderUnit` field in `operation.Receivers`.

After this patch, `app/leader` to run a task on the leader unit of an application will be supported
by both `juju run` and `juju exec --unit`.

- Refactored `makeOperationReceivers` to differentiate regular units from leader units.
- Moved `leader` constant to a shared context