refactor: rename types to be more obvious
Local and Remote are painful to read when context really matters. What
is local and to whom, and the same for remote. This change makes it a
lot more obvious. Consumer vs Offerer is always the same no matter the
context.

More of this will change as we go on, but for now make the simple
changes.