fix(modelmanager): replace BlockChecker interface usage
This patch fixes an issue in modelmanager, which causes the wrong model being checked for removal
and changes block. Before this patch, the controller model was checked for everything, so that it
was not possible to prevent the destroy or config changes for a specific model (it would always
check the controller model block status. After this patch, the right model is checked for both
changes and destroy, and `juju disable-command` will works as expected for destroy and changes.