fix(tests/cli): update base OS versions for deployments
This patch updates the base OS version for the `ubuntu-lite` and `ntp` deployments
from `ubuntu@20.04` to `ubuntu@24.04` in `tests/suites/cli/block.sh`
for compatibility and to align with the new default versions.

Not specifying the base for both ubuntu and ntp caused a CLI break since they need to have the same
one.