feat: handle the dying relation
If the consumer relation is dying, then we try our best to tell the
offerer side that the relation is going away. EXCEPT that if we get
a discharge error we can't do anything to notify the offerer of the
relation is gone. If the worker restarts the relation will likely
return RelationNotFound or we just don't get the event. This is
because relations do not transition to dead. Instead, they only go to
dying and then are removed. We get a one shot at this!

The offerer side will now be in a hung state. The solution is to be
able to remove an invisible offer via a relation ID which the operator
will have to do a sql query for (we'll improve ergonmics as we go).