feat: offer status watcher
Implement a watcher that emits notifications when an offer's
status changes

This is quite intricate, because many different things can lead to an
offer's status changing.

Namely, any change to the status of the underlying application's status,
which itself is calculated from the many statuses of all it's units.

But also, we need to watch for changes to the offer itself. Fortunately,
the only way an offer can change is termination. So we emit a
notification when the offer is removed.