feat: add new default storage pool uuid getter
This PR introduces a new func to the storage domain for getting the
default storage pool uuid and or making a new uuid.

This is required for recommended storage pool uuids. We want storage
pools that are recommended and are the same as the default pools to use
the same uuid.

If the recommended pool is not a default pool then a new uuid must be
created.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>