feat: set model recommended storage pools
This commit adds the ability for adding recommended storage pools to the
model when seeding the set of default storage pools. This works by
querying the storage registry to ask what pool it would recommend for a
given storage kind.

If the registry decides to provide a recommendation this is then saved
back down into the model to augment application deployment when selcting
a storage pool to use.

While I hear I broke out storage related testing into its own test
suite. This makes it easier to consume and work with.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>