feat: simplify report channel
The report channel didn't need to be so complicated, as we can just
wait for the consumer to read the channel without the need to pass
a channel to pass back.

We can expect the sub workers to disapear without notice, so check for
a timeout and if the catacomb is dying. Thus we can return back quickly
without waiting for the timeout - which might seem like a lifetime.