refactor: introduce modified model storage pool getter
Before this change we had the concept of getting storage pool
provisioners from state for the application domain. They were originally
called provisioners because they could be either a storage pool uuid or
a provider type.

We did away with the provider type idiom and so it now makes sense to
just talk about things with respect to getting storage pools for
specific tasks.

A recent commit introduced the concept of model default storage pools.
This commit builds on top of this by making the state layer consider
these defaults and that which has been supplied in model config.

Model config will always overwrite the model defaults in this case. We
except over time these model config keys will be removed and it will
just be the table that gets talks about in the state layer.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>