refactor: lxd provider recommended storage pool
This change refactors the lxd provider to now return the correct
recommended storage pool for filesystems out of it's registry.

Previous to this the value was being overriden via model config. As we
are trying to move away from model config for this type of configuration
it makes sense for this to be returned as a recommended storage pool.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>