refactor: openstack recommend storage pool
Refactors the openstack recommended storage pool to only return the
cinder pool when it is available to the environ. If it isn't available
it should only return one of the IAAS builting pools.

We were also incorrectly setting the storge to cinder in model config
even when it wasn't available.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>