feat: wire up offerer relation changes
If the offerer unit or relation changes we need to know. Thus take the
information and pass it on to the cross model relation service to deal
with.

Lastly, we need to know what units are in scope so we don't accidently
kill off something before everything has left the relation.