refactor: rename default storage pool method
Recently I added a new func to either return the well known uuid for a
storage pool or generate a new one if it wasn't available. It was
semantically similar naming to the version that will error if the
default doesn't exist.

PR feedback suggested a larger name change to seperate them from each
other.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>