refactor: better establishment of cinder storage in openstack provider
Before this commit within the registry provider of openstack we were
constantly checking if the environ region supported cinder storage. This
code path was repetitive and also introduce a line of errors that don't
need to exist.

By having the environ establish the facts around cinder on open much
like what is done with networking we can keep referring back to this
state during the life of the environ.

Having this value set also reduces contention on the shared mutex of the
provider.

As part of this change I updated unit tests and introduced missing unit
tests to show the code is predictible and operates to the contract.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>