fix: logic error when setting model storage pools
This fixes a logic error when setting a model's default storage pools.
If the storage pool uuid was use multiple times as a default for
different storage kinds then it would fail the check to ensure the pool
uuids existed. This is because the count from sql was deduplicated but
the slice was not.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>