feat: remove remote relation offerer
There is no need to have the remoterelationofferer worker in the offerer
side. The purpose of the consumer one is to consume and reflect the
changes. The offerer is already dispatching the changes via the
watchers, which means it's already doing the job. The reason we can
achieve this is because in 3.6 there is no way (easily) to know which
side you're operating on, so you need to witness all remote
applications. In 4.0 we do know this, because we insert a different
document for each side; offerer and consumer ones. Thus we only need
to watch the consumer creation.