docs: update test to include a description of the behaviour under test
Added a doc string for a storage provisioning watcher test so that I
could understand what the test that is being flaky is trying to assert.

The values being modified to observe change in the test did not directly
correlate to the watcher. This made it even harder to understand.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>