test: change storage attachment life in one transaction
This does the following two things:

1. Change the life of a storage attachment using the attachment uuid.
   This is the primary key of the record under modification and so it
   makes sense to use this values. Previous to this we were using the
   storage instance uuid which makes no sense and confuses what the test
   is doing.

2. Perform the storage attachment life changes in one transaction.
   Because this test changes the life of multiple attachments per
   harness test it was highlighting a bug in the event multiplexer
   subscription.

This is because not all attachment changes would come out in a single
term. When the changes get split over multiple terms we end up with a
dead lock on the subscription because it requires the watcher caller to
consume the available changes before sending a new set.

The harness would fail because it needs to see the change stream become
idle which can only happen once all changes have been dispatched.

With the way the watcher harness works the changes cannot be consumed
until the change stream becomes idle. This places the test in a chicken
and egg scenario causing the test context to eventually fail.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>