fix: fixes a bug that does not allow the user to override the storage directive count
A bug was noticed during testing that when a user asks for a storage
count override but the charm has not specified a maximum we reject this
and provide an error to the user.

This isn't correct and should be allowed. To help test the bug fix and
also communicate better to the user a new error type has been introduced
to help the apiserver provide more context downwards.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>