feat: ensure we can set empty settings
Empty settings for both the units and applications are allowed to clear
the settings of a relation. If the maps are nil, then updating them
should be skipped.

Lastly, the setting of application and unit settings was overloaded
to support nil application settings. This could cause issues if you
didn't know what nil vs empty was. Just split the methods up and
have a better clean API.