feat: remote relation app and unit settings
We want to set the remote relation application and unit settings. This
allows us to update the values from the offerer side, so they're
replicated.

We don't need to use the leadership service here, as we're forcing
the values from the offerer. That leadership check was done by the
original controller (i.e. it wouldn't have been set unless they where
the leader).