refactor: rename struct elements to suit the data
RelationUnitChanges returns all of the current data after a the watcher
fires indicating something has changed. ChangedUnits was misnamed as
it contains all of the units settings, the watcher won't indicated which
one has changed.

Updated comments to indicate why IDs are used, rather than unit names.
The application on the other end of the relation may have a different
name from what it is locally.