refactor: change the name of the func for checking if a unit is registered
This commit changes the name of the func that is used to check if a unit
is registered in the model. The name was previously prefixed with Check
and not is prefixed with Get. This better matches the intent of what the
function is doing.

Done as part of PR feedback.

Signed-off-by: Thomas Miller <thomas.miller@canonical.com>