feat(status): add remote application (SAAS) to status
This required a new service method to bulk get the statuses of remote
application offerers.

For this, I needed to include offer urls in the db. Refactor
"AddRemoteApplicationOfferer" to store this piece of data as well.

Wire this up to the FullStatus facade.

As a flyby, refactor usage of offer url to use a concrete type instead of
a pointer